DSCP were not matched correctly, also DSCP VA was missing (10.0 version)
DSCP values specified in a firewall rule matched not only the specified value, but also other values. See https://forum.pfsense.org/index.php/topic,67824.0.html

Also, the VA code point was missing, so i added it while I was on it.

This patch is equivalent to my patch for 8.3 version.